feat: fix tweet function
